Change hash method to use unsigned long

+++ cploration/c07/symtable.c
@@ -10,7 +10,7 @@
 Symbol* hashArray[SYMBOL_TABLE_SIZE];
 
 int hash(char *string) {
-    int hash = 5381;
+    unsigned long hash = 5381;
     int c;
 
     while ((c = *string++)) {
